Add CI using cross (#24)
* test cross build CI

* bump to aws-lc-sys version that supports cross build

* run all tests on all three cross platforms

* fix digest test failure on 32 bit platforms

* use u64 as the input limit for larger hash blocks

+++ src/digest/sha.rs
@@ -42,11 +42,12 @@ const SHA512_BLOCK_LEN: usize = 1024 / 8;
 
 /// SHA-1 and SHA-256 are limited to an input size of 2^64-1 bits.
 #[allow(clippy::cast_possible_truncation)]
-const SHA256_MAX_INPUT_LEN: usize = u64::MAX as usize;
+const SHA256_MAX_INPUT_LEN: u64 = u64::MAX;
 
-/// SHA-384, SHA-512, and SHA-512/256 are limited to an input size of 2^128-1 bits.
+/// SHA-384, SHA-512, and SHA-512/256 are limited to an input size of 2^128-1 bits according to the spec.
+/// u64 is more than sufficient enough for practical usecases, so we limit the input length to 2^64-1 bits.
 #[allow(clippy::cast_possible_truncation)]
-const SHA512_MAX_INPUT_LEN: usize = u128::MAX as usize;
+const SHA512_MAX_INPUT_LEN: u64 = u64::MAX;
 
 /// SHA-1 as specified in [FIPS 180-4]. Deprecated.
 ///
